# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GamePlatform = typeof GamePlatform[keyof typeof GamePlatform];


export const GamePlatform = {
  PC: 'PC',
  Xbox: 'Xbox',
  PlayStation: 'PlayStation',
} as const;

export type GameSourceStatus = typeof GameSourceStatus[keyof typeof GameSourceStatus];


export const GameSourceStatus = {
  live: 'live',
  snapshot: 'snapshot',
  unavailable: 'unavailable',
} as const;

export interface Game {
  id: string;
  name: string;
  platform: GamePlatform;
  provider: string;
  genre: string[];
  imageUrl: string;
  storeUrl: string;
  reviewUrl: string;
  currency: string;
  currentPrice: number;
  originalPrice: number;
  discountPercent: number;
  ratingOutOfFive: number;
  reviewCount: number;
  dopeScore: number;
}

export interface GameSource {
  name: string;
  url: string;
  status: GameSourceStatus;
  detail: string;
}

export interface GameCatalog {
  platform: GamePlatform;
  region: string;
  currency: string;
  games: Game[];
  sources: GameSource[];
  /** @nullable */
  refreshedAt: string | null;
  message: string;
}

export interface GameCatalogSummary {
  platform: GamePlatform;
  gameCount: number;
  topDopeScore: number;
  genres: string[];
  sourceStatus: GameSourceStatus;
}

export type GetGameCatalogParams = {
platform: GamePlatform;
genre?: string;
search?: string;
/**
 * Two-letter storefront country code
 * @minLength 2
 * @maxLength 2
 */
region?: string;
};

export type GetGameCatalogSummaryParams = {
platform: GamePlatform;
};