# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: games
    description: Public game listings and value scores
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /games:
    get:
      operationId: getGameCatalog
      tags: [games]
      summary: Get value-ranked games for a platform
      parameters:
        - name: platform
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/GamePlatform"
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: region
          in: query
          required: false
          description: Two-letter storefront country code
          schema:
            type: string
            minLength: 2
            maxLength: 2
      responses:
        "200":
          description: Game catalog and data-source status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameCatalog"
  /games/summary:
    get:
      operationId: getGameCatalogSummary
      tags: [games]
      summary: Get catalog counts and available genres
      parameters:
        - name: platform
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/GamePlatform"
      responses:
        "200":
          description: Summary of games available for the selected platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameCatalogSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    GamePlatform:
      type: string
      enum: [PC, Xbox, PlayStation]
    GameSourceStatus:
      type: string
      enum: [live, snapshot, unavailable]
    Game:
      type: object
      required:
        - id
        - name
        - platform
        - provider
        - genre
        - imageUrl
        - storeUrl
        - reviewUrl
        - currency
        - currentPrice
        - originalPrice
        - discountPercent
        - ratingOutOfFive
        - reviewCount
        - dopeScore
      properties:
        id: { type: string }
        name: { type: string }
        platform: { $ref: "#/components/schemas/GamePlatform" }
        provider: { type: string }
        genre: { type: array, items: { type: string } }
        imageUrl: { type: string }
        storeUrl: { type: string }
        reviewUrl: { type: string }
        currency: { type: string }
        currentPrice: { type: number }
        originalPrice: { type: number }
        discountPercent: { type: number }
        ratingOutOfFive: { type: number }
        reviewCount: { type: integer }
        dopeScore: { type: number }
    GameSource:
      type: object
      required: [name, url, status, detail]
      properties:
        name: { type: string }
        url: { type: string }
        status: { $ref: "#/components/schemas/GameSourceStatus" }
        detail: { type: string }
    GameCatalog:
      type: object
      required: [platform, region, currency, games, sources, refreshedAt, message]
      properties:
        platform: { $ref: "#/components/schemas/GamePlatform" }
        region: { type: string }
        currency: { type: string }
        games:
          type: array
          items: { $ref: "#/components/schemas/Game" }
        sources:
          type: array
          items: { $ref: "#/components/schemas/GameSource" }
        refreshedAt: { type: ["string", "null"], format: date-time }
        message: { type: string }
    GameCatalogSummary:
      type: object
      required: [platform, gameCount, topDopeScore, genres, sourceStatus]
      properties:
        platform: { $ref: "#/components/schemas/GamePlatform" }
        gameCount: { type: integer }
        topDopeScore: { type: number }
        genres:
          type: array
          items: { type: string }
        sourceStatus: { $ref: "#/components/schemas/GameSourceStatus" }