# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GamePlatform } from './gamePlatform';

export interface Game {
  id: string;
  name: string;
  platform: GamePlatform;
  provider: string;
  genre: string[];
  imageUrl: string;
  storeUrl: string;
  reviewUrl: string;
  currency: string;
  currentPrice: number;
  originalPrice: number;
  discountPercent: number;
  ratingOutOfFive: number;
  reviewCount: number;
  dopeScore: number;
}