# lib/api-zod/src/generated/types/gameCatalog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Game } from './game';
import type { GamePlatform } from './gamePlatform';
import type { GameSource } from './gameSource';

export interface GameCatalog {
  platform: GamePlatform;
  region: string;
  currency: string;
  games: Game[];
  sources: GameSource[];
  /** @nullable */
  refreshedAt: Date | null;
  message: string;
}