# lib/api-zod/src/generated/types/gameCatalogSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GamePlatform } from './gamePlatform';
import type { GameSourceStatus } from './gameSourceStatus';

export interface GameCatalogSummary {
  platform: GamePlatform;
  gameCount: number;
  topDopeScore: number;
  genres: string[];
  sourceStatus: GameSourceStatus;
}