# lib/api-zod/src/generated/types/gamePlatform.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type GamePlatform = typeof GamePlatform[keyof typeof GamePlatform];


export const GamePlatform = {
  PC: 'PC',
  Xbox: 'Xbox',
  PlayStation: 'PlayStation',
} as const;