# lib/api-zod/src/generated/types/gameSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameSourceStatus } from './gameSourceStatus';

export interface GameSource {
  name: string;
  url: string;
  status: GameSourceStatus;
  detail: string;
}