# lib/api-zod/src/generated/types/gameSourceStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type GameSourceStatus = typeof GameSourceStatus[keyof typeof GameSourceStatus];


export const GameSourceStatus = {
  live: 'live',
  snapshot: 'snapshot',
  unavailable: 'unavailable',
} as const;